:lipstick: :: spacing 값 수정

+++ lib/presentation/map_page/widget/place_review_container.dart
@@ -97,7 +97,7 @@ class PlaceReviewContainer extends StatelessWidget {
               ),
             ),
             Padding(
-              padding: const EdgeInsets.only(left: AppSpacing.s16),
+              padding: const EdgeInsets.only(left: AppSpacing.s8),
               child: IconButton(onPressed: () {}, icon: AppIcons.bin()),
             ),
           ],
